+++ src/Pdf++.Core/include/CPPPdf/Document.hpp
@@ -6,3 +6,4 @@
 #include <CPPPdf/Document/PdfPageOrganizer.hpp>
 #include <CPPPdf/Document/PdfPageImporter.hpp>
 #include <CPPPdf/Document/PdfDocumentLayout.hpp>
+#include <CPPPdf/Document/PdfRedactor.hpp>